Moved AppBinder and Main to the app package
