+++ pom.xml
@@ -248,7 +248,7 @@
                 <configuration>
                     <archive>
                         <manifest>
-                            <mainClass>org.fsg1.fmms.backend.resources.Main</mainClass>
+                            <mainClass>org.fsg1.fmms.backend.app.Main</mainClass>
                             <addClasspath>true</addClasspath>
                         </manifest>
                     </archive>
